// src/types/games.d.ts
export type Game = {
  _id: string;
  Name: string;
  "Release date": string;
  Price: number;
  "About the game": string;
  "Supported languages": string;
  "Header image": string;
  Developers: string;
  Publishers: string;
  Categories: string;
  Genres: string;
  Tags: string | null;
  Screenshots: string;
  Movies: string;
};